requirements.txt
# Sistema NPS Automation - Dependências para Windows
# Versões específicas testadas para compatibilidade

# Framework Web
flask==2.3.3
flask-cors==4.0.0

# HTTP Requests
requests==2.31.0

# OpenAI API
openai==1.3.0

# Configuração
python-dotenv==1.0.0

# Terminal/Console (Windows)
colorama==0.4.6

# Documentos
python-docx==0.8.11

# Planilhas
openpyxl==3.1.2
pandas==2.0.3

# PDF Generation (WeasyPrint com dependências Windows)
weasyprint==60.2
cffi==1.16.0
cairocffi==1.6.1
cairosvg==2.7.1

# Windows específico
pywin32==306; sys_platform == "win32"

# Unicode support
unicodedata2==15.1.0

# Cache e performance
lru-dict==1.2.0
cachetools==5.3.2